requirements.txt
Django==3.1.7
django-jazzmin==2.2.5
django-tinymce==3.3.0
pytils==0.3
django-multiselectfield==0.1.12
pillow==8.3.1
psycopg2-binary